requirements.txt
PyMuPDF
Pillow
pandas
openpyxl
